requirements-importer.txt
requests
beautifulsoup4
lxml
tqdm
python-dotenv